engine/requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-chroma>=0.0.1
langchain-core>=0.1.0  # 添加 core 依赖
pypdf>=3.0.0  # 将 PyPDF2>=3.0.1 改为pypdf
pydantic>=2.0.0
aiohttp>=3.8.0
requests>=2.31.0
python-docx>=0.8.11
pandas>=2.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0